[FEATURE] add pagination getAllBooking

+++ src/main/java/com/enigma/services/impl/BookingServiceImpl.java
@@ -11,6 +11,8 @@ import com.enigma.services.EventService;
 import com.enigma.services.TicketService;
 import com.enigma.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
@@ -30,27 +32,12 @@ public class BookingServiceImpl implements BookingService {
 
     @Override
     public Booking booking(Booking booking) {
-        User user = userService.getUserById(booking.getUserIdTransient());
-        booking.setUser(user);
-        booking.setPaymentStatus(false);
-        BigDecimal sumSubtotal;
-
-        for (BookingDetail bookingDetail : booking.getBookingDetailList()) {
-            bookingDetail.setBooking(booking);
-            Ticket ticket = ticketService.getTicketById(bookingDetail.getTicketIdTransient());
-            bookingDetail.setTicket(ticket);
-
-            ticketService.deduct(bookingDetail.getTicket().getId(), bookingDetail.getQuantity());
-            sumSubtotal = ticket.getPrice().multiply(new BigDecimal(bookingDetail.getQuantity()));
-            bookingDetail.setSubtotal(sumSubtotal);
-
-        }
         return bookingRepository.save(booking);
     }
 
     @Override
-    public List<Booking> getAllBookingData() {
-        return bookingRepository.findAll();
+    public Page<Booking> getAllBookingData(Pageable pageable) {
+        return bookingRepository.findAll(pageable);
     }
 
     @Override
@@ -63,11 +50,8 @@ public class BookingServiceImpl implements BookingService {
 
     @Override
     public void deleteBookingDataById(String bookingId) {
-        Booking booking = getBookingById(bookingId);
-        for (BookingDetail bookingDetail : booking.getBookingDetailList()) {
-            String idTicket = bookingDetail.getTicket().getId();
-            Ticket ticket = ticketService.getTicketById(idTicket);
-            ticket.setQuantity(ticket.getQuantity() + bookingDetail.getQuantity());
+        if (!bookingRepository.findById(bookingId).isPresent()) {
+            throw new NotFoundException(String.format(MessageConstant.ID_BOOKING_NOT_FOUND, bookingId));
         }
         bookingRepository.deleteById(bookingId);
     }
